SkillCanceller::Wait/StopWait
This adds a third option how SkillCanceller can be used.

Why?
The lifetime of the references that the callback uses are not long enough.
Any useful callback needs to capture a few references to do the
cancellation (typically icon::Session), but these references don't live long enough,
not as long as the SkillCanceller.
This fixes the lifetime issue by not using callbacks in the skills.
On the skill side, it is slightly more involved to use,
but it is now memory safe and thread safe.

GitOrigin-RevId: 7ed13682ac13c9b53cb11d79bd0e3844b3b60c52